bluetooth: Standalone Tests for characteristic/
This change refactors the Web Bluetooth WPT tests in the characteristic/
directory to be standalone window tests, as well as use the async/await
pattern. This makes the tests easier to format and easier to read.

The generated tests are not updated in this patch because they will
require updating the Python script that generates them.

Bug: 509038
Change-Id: I74d2158bbff3ef3b30d5286129cc3fc855e9a9da