Include initial frame replication state on portal adoption.
This is needed so that checks that occur as part of navigation
don't get confused by the replication state (notably the origin)
of the portal guest contents being missing from the remote frame
in the host contents' renderer.

Since this state is pushed down into the blink::RemoteFrame,
we need InitializeCoreFrame to be called before we can set the
replication state. As a result WebRemoteFrame::CreateForAdoptedPortal
is created, which performs this initialization with the provided
HTMLPortalElement.

Bug: 980937
Change-Id: I843192e5f5aa8e4e022a88d318fa4a5e1d46a3d9