[Client hints] Enable ACHL persistence in content shell, refactor tests
Currently, Accept-CH-Lifetime tests are not covered by the bots, as
content_shell had no persistency for them. Moving the CH logic to
content/ [1] was a first step in that direction, but I never went ahead
with the rest of it and enabled persistency in content_shell.
This CL fixes that! It also rewrites the tests as they were no longer
passing.

Change-Id: I9c19bc52afd9ced5477da99ccce2e27ecc5d89c5

Bug:981027

[1] https://chromium-review.googlesource.com/c/1450127

Change-Id: I9c19bc52afd9ced5477da99ccce2e27ecc5d89c5